#125 fix clang compilation -Wfloat-conversion

+++ dynawo/sources/Solvers/AlgebraicSolvers/DYNSolverKINCommon.cpp
@@ -137,7 +137,7 @@ SolverKINCommon::initCommon(const std::string& linearSolverName, double fnormtol
   // -------------------
   // Passing CSR_MAT indicates that we solve A'x = B - linear system using the matrix transpose -
   // and not Ax = B (see sunlinsol_klu.c:149)
-  const int nnz = 0.0001 * numF_ * numF_;  // This value will be adjusted later on in the process
+  const int nnz = static_cast<int>(0.0001 * numF_ * numF_);  // This value will be adjusted later on in the process
   sundialsMatrix_ = SUNSparseMatrix(numF_, numF_, nnz, CSR_MAT);
   if (sundialsMatrix_ == NULL)
     throw DYNError(Error::SUNDIALS_ERROR, SolverFuncErrorKINSOL, "SUNSparseMatrix");
